fix: iOS 앱 출시 전 버그 수정
- iOS 기기별 OverFlowedPixel 에러 해결

- HomePage에서 답장할 게시글의 Doc ID를 잘못 전달하는 에러 해결

- Firestore ChatModel 구조로 인해 발생하는 에러 해결

+++ lib/presentation/pages/login/login_view_model.dart
@@ -1,7 +1,6 @@
 // ignore_for_file: file_names, avoid_print
 
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:everyones_tone/app/repository/firestore_data.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 import 'package:sign_in_with_apple/sign_in_with_apple.dart';
@@ -67,26 +66,33 @@ class LoginViewModel {
     await _googleSignIn.signOut();
   }
 
-  //! 회원탈퇴
   Future<void> deleteUserAccount() async {
     try {
-      // 일단 로그아웃하기
-      await _auth.signOut();
-      await _googleSignIn.signOut();
+      final User? currentUser = _auth.currentUser;
+      if (currentUser == null) {
+        throw Exception("User is not logged in");
+      }
 
-      var userDocRef =
-          _firestore.collection('user').doc(FirestoreData.currentUserEmail);
+      var userDocRef = _firestore.collection('user').doc(currentUser.email);
 
-      // 서브컬렉션 내의 모든 문서를 가져옵니다.
       var myChatSnapshot = await userDocRef.collection('myChat').get();
       for (var myChat in myChatSnapshot.docs) {
         await myChat.reference.delete();
       }
 
-      // 채팅방 문서 자체를 삭제합니다.
+      var previousRepliesSnapshot =
+          await userDocRef.collection('previousReplies').get();
+      for (var previousReplies in previousRepliesSnapshot.docs) {
+        await previousReplies.reference.delete();
+      }
+
+      // 채팅방 문서 자체를 삭제
       await userDocRef.delete();
+
+      // 로그아웃하기
+      await _auth.signOut();
+      await _googleSignIn.signOut();
     } catch (e) {
-      // 오류 처리
       print("회원 탈퇴 중 오류 발생: $e");
     }
   }
